+++ src/index.ts
@@ -1,3 +1,4 @@
+// 1. Exports plugin for eslint to use
 /**
  * @author Amila Welihinda
  */
