--- src/providers/kangax-provider.ts
@@ -1,6 +1,0 @@
-import { Node } from "../types";
-
-// TODO: not implemented yet.
-const KangaxProvider: Array<Node> = [];
-
-export default KangaxProvider;
